New UI event type for BUTTON_DOWN
Enables actions to fire immediately when pressing a button, instead of when you release it.
Existing functionality for BUTTON_PRESS and BUTTON_LONG_PRESS works the same.
I tried to fix all naive event.type assumptions in legacy apps.
Hemisphere applets now use the new event type, which might be weird in
some cases, but we're gonna try it!