FreqMeasure: fix end() on T4.x; different pins on T4.0 vs T4.1; cleanup
T4.0 works fine for TR1/TR2 (pin 0/1) but not for TR3/TR4 (pin 2/3)
It will work for pin 23/22 but those are connected to the encoders on
original hardware. I imagine the driver code needs different parameters
for FlexPWM to work on pins 2/3 instead (or maybe QuadTimer will also
work on pins 2/3), but I can't swim through the data sheets to figure it out!